requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.2
django-allauth==0.48.0
django-betterforms==1.2
django-countries==7.2.1
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-extensions==3.1.5
django-heroku==0.3.1
django-werkzeug==1.0.0
djangorestframework==3.13.1
djlint==0.7.4
environ==1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
mysqlclient==2.1.0
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
razorpay==1.3.0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
stripe==2.66.0
toml==0.10.2
tomlkit==0.8.0
tqdm==4.62.3
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.3
whitenoise==6.0.0
zipp==3.7.0
